Added DotsChecker. Fixed some warnings

+++ backend/src/main/java/com/forafox/web_lab_4/models/token/RefreshTokenRepository.java
@@ -15,5 +15,5 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
     @Modifying
     int deleteByUser(User user);
 
-    Optional<RefreshToken> getByUserId(Integer id);
+    Optional<RefreshToken> getByUserId(Long id);
 }
